services/rag-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
qdrant-client==1.7.0
langchain==0.0.350
langchain-community==0.0.10
langchain-core>=0.1.8
sentence-transformers==2.7.0
huggingface-hub==0.24.6
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
aiofiles==23.2.1
minio==7.2.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic==1.13.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
